requirements.txt
numpy==2.0.1
openai==1.40.3
opencv_python==4.10.0.84
Pillow==10.4.0
PyAutoGUI==0.9.54
python-dotenv==1.0.1
replicate==0.31.0
Requests==2.32.3
